fix(ci): restore restart health lint
